Create IgnitionWindowComponentDict
Recursive function that crawls down a window's component tree and adds the components name (key) and object to a dictionary which is then stored in the global dictionary. Can then be used to reference components with just a name without walking up and down .parent.parent.getComponent to find components and having script break if you move something in or out of a container.
This is the initial proof of concept (and some commented out development versions). Was developed on a button. Probably better as a window initialization script.